api: add unknown module
This commit adds an "unknown" module as a catch-all for
unsupported platforms (non-Unix, non-Windows). This is
useful for platforms like WASM that don't support these
primitives.

While one defensible approach would be the status quo,
this would require dependents to treat `same-file` (and all
of its dependents) as platform dependent dependencies,
which is a bit inconvenient.

Instead, the approach we take here is to mirror the public
API for unsupported platforms, but return an error for all
accesses. This will allow programs to compile, but
potentially result in runtime errors. The WASM folks seem
to be okay with this, so we roll with it.

The only real concern here is that adding this in a release
means we are committing to it. So if this turns out to be a
mistake, we can't remove it without a major version bump.
The risk here seems pretty small.